requirements.txt
pandas>=2.0.0
openai>=1.0.0
chromadb>=0.4.0
tqdm>=4.65.0
python-dotenv>=1.0.0
numpy>=1.24.0
streamlit>=1.28.0
